+++ webpack.config.js
@@ -32,10 +32,6 @@ module.exports = {
         {
           from: 'node_modules/file-icons-js/fonts/',
           to: 'vendors/file-icons-js/fonts/'
-        },
-        {
-          from: 'node_modules/jquery/dist/jquery.slim.min.js',
-          to: 'vendors/jquery.slim.min.js'
         }
       ]
     })
